Creates test file for shared state/action
