+++ kaskade/coroutines/src/commonTest/kotlin/dev/gumil/kaskade/coroutines/test.kt
@@ -1,0 +1,17 @@
+package dev.gumil.kaskade.coroutines
+
+import dev.gumil.kaskade.Action
+import dev.gumil.kaskade.SingleEvent
+import dev.gumil.kaskade.State
+
+sealed class TestState : State {
+    object State1 : TestState()
+    object State2 : TestState()
+    object SingleStateEvent : TestState(), SingleEvent
+}
+
+sealed class TestAction : Action {
+    object Action1 : TestAction()
+    object Action2 : TestAction()
+    object Action3 : TestAction()
+}
